ボーナスゲーム処理仮実装 / フラグ処理修正
1:ボーナスゲームの処理を実装

2:フラグ処理を改良

+++ Assets/Scripts/Reel/ReelTableManager.cs
@@ -80,7 +80,7 @@ public class ReelTableManager
         }
         // 見つけたリールテーブルを記録
         Debug.Log("Final Found:" + foundTable);
-        UsedReelTableID[reel.ReelID] = foundTable;
+        UsedReelTableID[(int)reel.ReelID] = foundTable;
         return foundTable;
     }
